Addressed Le's comments
