+++ src/routers/admin.py
@@ -9,8 +9,11 @@ from entities.models import UserProfile
 
 from regtech_api_commons.models.auth import AuthenticatedUser
 from regtech_api_commons.oauth2.oauth2_admin import OAuth2Admin
+from config import kc_settings
 
 router = Router()
+
+oauth2_admin = OAuth2Admin(kc_settings)
 
 
 @router.get("/me/", response_model=AuthenticatedUser)
@@ -22,12 +25,12 @@ def get_me(request: Request):
 @router.put("/me/", status_code=HTTPStatus.ACCEPTED, dependencies=[Depends(check_domain)])
 @requires("manage-account")
 def update_me(request: Request, user: UserProfile):
-    OAuth2Admin.update_user(request.user.id, user.to_keycloak_user())
+    oauth2_admin.update_user(request.user.id, user.to_keycloak_user())
     if user.leis:
-        OAuth2Admin.associate_to_leis(request.user.id, user.leis)
+        oauth2_admin.associate_to_leis(request.user.id, user.leis)
 
 
 @router.put("/me/institutions/", status_code=HTTPStatus.ACCEPTED, dependencies=[Depends(check_domain)])
 @requires("manage-account")
 def associate_lei(request: Request, leis: Set[str]):
-    OAuth2Admin.associate_to_leis(request.user.id, leis)
+    oauth2_admin.associate_to_leis(request.user.id, leis)
